+++ examples/kubeflow/pipeline/clean_text/.s2i/environment
@@ -1,0 +1,4 @@
+MODEL_NAME=RedditClassifier
+API_TYPE=REST
+SERVICE_TYPE=MODEL
+PERSISTENCE=0
